code/pyproject.toml
[tool.poetry]
name = "plantapop"
version = "0.1.0"
description = "El núcleo de Plantacity, una aplicación para la venta e intercambio de plantas."
authors = ["pablo-snz <pablo.snz.brm@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
SQLAlchemy = {version = "^2.0.18", extras = ["asyncio"]}
alembic = "^1.11.1"
pydantic = "^2.0.2"
PyYAML = "^6.0"
pydantic-settings = "^2.0.2"
sqlalchemy-utils = "^0.41.1"
omegaconf = "^2.3.0"
httpx = "^0.25.1"
dependency-injector = "^4.41.0"
toml = "^0.10.2"
langcodes = "^3.3.0"
pytz = "^2023.3.post1"
bcrypt = "^4.0.1"
sqlalchemy-filters = "^0.13.0"
freezegun = "^1.2.2"
aio-pika = "^9.3.1"
aiosqlite = "^0.19.0"
asyncpg = "^0.29.0"
redis = "^5.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
python-jose = "^3.3.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
ruff = "^0.1.5"
pyproject-flake8 = "^6.1.0"
debugpy = "^1.8.0"

[tool.poetry.scripts]
start = "plantapop.__main__:main"
unittest = "ci:run_unittest"
integration = "ci:run_integrationtest"
acceptance = "ci:run_acceptancetest"
format = "ci:run_format"
lint = "ci:run_lint"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
ignore = ["W292", "W503", "E203"]

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["**/migrations/**"]
